# app/Http/Controllers/Api/SearchController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;
use App\Models\Advert;
use App\Models\SubCategory;
use App\Models\Category;
use App\Models\Brands;
use App\Models\State;
use Jenssegers\Agent\Agent;

/**
 * @group Search
 *
 * APIs for searching and filtering adverts across the platform
 */
class SearchController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/search",
     *     summary="Search adverts with filters",
     *     tags={"Search"},
     *     @OA\Parameter(name="product", in="query", description="Product search term", @OA\Schema(type="string")),
     *     @OA\Parameter(name="location", in="query", description="State/Location", @OA\Schema(type="string")),
     *     @OA\Parameter(name="category", in="query", description="Category ID", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="sub_category", in="query", description="Sub-category ID", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="brand", in="query", description="Brand ID", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="buydirect", in="query", description="Buy direct filter", @OA\Schema(type="string")),
     *     @OA\Parameter(name="page", in="query", description="Page number", @OA\Schema(type="integer")),
     *     @OA\Response(response=200, description="Search results")
     * )
     */
    public function search(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'product' => 'nullable|string|min:3',
            'location' => 'nullable',
            'category' => 'nullable',
            'sub_category' => 'nullable',
            'brand' => 'nullable',
            'buydirect' => 'nullable',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'errors' => $validator->errors()
            ], 422);
        }

        $query = Advert::with('firstImage')->activeNotRecentlySold();

        if ($request->filled('product')) {
            $query->where(function ($q) use ($request) {
                $q->where('ad_title', 'LIKE', '%' . $request->product . '%')
                  ->orWhere('ad_id', 'LIKE', '%' . $request->product . '%');
            });
        }

        if ($request->filled('category')) {
            $query->where('category', $request->category);
        }

        if ($request->filled('sub_category')) {
            $query->where('sub_category', $request->sub_category);
        }

        if ($request->filled('brand')) {
            $query->where('brand', $request->brand);
        }

        if ($request->filled('location')) {
            $query->where('state', $request->location);
        }

        if ($request->filled('buydirect')) {
            $query->where('buy_direct', $request->buydirect);
        }

        $ads = $query->orderWithFeatured()
                    ->paginate($request->input('per_page', 10));

        return response()->json([
            'success' => true,
            'data' => $ads->items(),
            'pagination' => [
                'current_page' => $ads->currentPage(),
                'last_page' => $ads->lastPage(),
                'per_page' => $ads->perPage(),
                'total' => $ads->total(),
                'has_more' => $ads->hasMorePages()
            ],
            'search_params' => [
                'product' => $request->input('product'),
                'location' => $request->input('location'),
                'category' => $request->input('category'),
                'sub_category' => $request->input('sub_category'),
                'brand' => $request->input('brand'),
                'buydirect' => $request->input('buydirect'),
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/search/filter",
     *     summary="Advanced filtering with price ranges",
     *     tags={"Search"},
     *     @OA\RequestBody(
     *         required=false,
     *         @OA\JsonContent(
     *             @OA\Property(property="category", type="integer"),
     *             @OA\Property(property="sub_category", type="integer"),
     *             @OA\Property(property="brand", type="integer"),
     *             @OA\Property(property="location", type="string"),
     *             @OA\Property(property="min", type="number"),
     *             @OA\Property(property="max", type="number"),
     *             @OA\Property(property="range", type="string", enum={"under_20k", "20k_120k", "120k_1m", "1m_10m", "above_10m"})
     *         )
     *     ),
     *     @OA\Response(response=200, description="Filtered results")
     * )
     */
    public function filter(Request $request)
    {
        $query = Advert::with('firstImage')
                    ->where('ad_status', 'active')
                    ->where('sold', 'No');

        if ($request->filled('category')) {
            $query->where('category', $request->category);
        }

        if ($request->filled('sub_category')) {
            $query->where('sub_category', $request->sub_category);
        }

        if ($request->filled('brand')) {
            $query->where('brand', $request->brand);
        }

        if ($request->filled('location')) {
            $query->where('state', $request->location);
        }

        // Price filters
        $min = $request->input('min');
        $max = $request->input('max');
        $range = $request->input('range');

        if ($min !== null) {
            $query->where('price', '>=', (int) $min);
        }

        if ($max !== null) {
            $query->where('price', '<=', (int) $max);
        }

        if ($range) {
            switch ($range) {
                case 'under_20k':
                    $query->where('price', '<', 20000);
                    break;
                case '20k_120k':
                    $query->whereBetween('price', [20000, 120000]);
                    break;
                case '120k_1m':
                    $query->whereBetween('price', [120000, 1000000]);
                    break;
                case '1m_10m':
                    $query->whereBetween('price', [1000000, 10000000]);
                    break;
                case 'above_10m':
                    $query->where('price', '>', 10000000);
                    break;
            }
        }

        $adverts = $query->orderWithFeatured()
                        ->paginate($request->input('per_page', 10));

        return response()->json([
            'success' => true,
            'data' => $adverts->items(),
            'pagination' => [
                'current_page' => $adverts->currentPage(),
                'last_page' => $adverts->lastPage(),
                'per_page' => $adverts->perPage(),
                'total' => $adverts->total(),
                'has_more' => $adverts->hasMorePages()
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/search/filter-by-seller",
     *     summary="Filter by verified/unverified sellers",
     *     tags={"Search"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="sellers", type="string", enum={"all", "yes", "no"}, description="Seller verification status")
     *         )
     *     ),
     *     @OA\Response(response=200, description="Filtered results")
     * )
     */
    public function filterBySeller(Request $request)
    {
        $query = Advert::with('firstImage')
            ->where('ad_status', 'active')
            ->where('sold', 'No');

        $sellers = $request->input('sellers', 'all');
        if ($sellers !== 'all') {
            $query->whereHas('owner', function ($q) use ($sellers) {
                $q->where('verified', $sellers);
            });
        }

        // Context filters
        if ($request->filled('category')) {
            $query->where('category', $request->category);
        }

        if ($request->filled('sub_category')) {
            $query->where('sub_category', $request->sub_category);
        }

        if ($request->filled('brand')) {
            $query->where('brand', $request->brand);
        }

        if ($request->filled('location')) {
            $query->where('state', $request->location);
        }

        $adverts = $query->orderWithFeatured()
                        ->paginate($request->input('per_page', 10));

        return response()->json([
            'success' => true,
            'data' => $adverts->items(),
            'pagination' => [
                'current_page' => $adverts->currentPage(),
                'last_page' => $adverts->lastPage(),
                'per_page' => $adverts->perPage(),
                'total' => $adverts->total(),
                'has_more' => $adverts->hasMorePages()
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/search/filter-by-buydirect",
     *     summary="Filter by buy direct availability",
     *     tags={"Search"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="buydirect", type="string", description="Buy direct filter value")
     *         )
     *     ),
     *     @OA\Response(response=200, description="Filtered results")
     * )
     */
    public function filterByBuydirect(Request $request)
    {
        $query = Advert::with('firstImage')
            ->where('ad_status', 'active')
            ->where('sold', 'No');

        if ($request->filled('buydirect')) {
            $query->where('buy_direct', $request->buydirect);
        }

        // Context filters
        if ($request->filled('category')) {
            $query->where('category', $request->category);
        }

        if ($request->filled('sub_category')) {
            $query->where('sub_category', $request->sub_category);
        }

        if ($request->filled('brand')) {
            $query->where('brand', $request->brand);
        }

        if ($request->filled('location')) {
            $query->where('state', $request->location);
        }

        $adverts = $query->orderWithFeatured()
                        ->paginate($request->input('per_page', 10));

        return response()->json([
            'success' => true,
            'data' => $adverts->items(),
            'pagination' => [
                'current_page' => $adverts->currentPage(),
                'last_page' => $adverts->lastPage(),
                'per_page' => $adverts->perPage(),
                'total' => $adverts->total(),
                'has_more' => $adverts->hasMorePages()
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/search/location/{location}/{slug}",
     *     summary="Search by location and category/subcategory/brand",
     *     tags={"Search"},
     *     @OA\Parameter(name="location", in="path", required=true, @OA\Schema(type="string")),
     *     @OA\Parameter(name="slug", in="path", required=true, @OA\Schema(type="string")),
     *     @OA\Response(response=200, description="Location-based search results")
     * )
     */
    public function locationSearch($location, $slug)
    {
        // Determine type based on slug existence
        if (Category::where('category_slug', $slug)->exists()) {
            return $this->locationCategory($location, $slug);
        } elseif (SubCategory::where('sub_cat_slug', $slug)->exists()) {
            return $this->locationSubcat($location, $slug);
        } elseif (Brands::where('brand_slug', $slug)->exists()) {
            return $this->locationBrand($location, $slug);
        }

        return response()->json([
            'success' => false,
            'message' => 'Resource not found'
        ], 404);
    }

    /**
     * Location + Category search
     */
    private function locationCategory($location, $slug)
    {
        $cat = Category::where('category_slug', $slug)->first();

        if (!$cat) {
            return response()->json([
                'success' => false,
                'message' => 'Category not found'
            ], 404);
        }

        $ads = Advert::with('firstImage')
                    ->activeNotRecentlySold()
                    ->where('state', $location)
                    ->where('category', $cat->id)
                    ->orderWithFeatured()
                    ->paginate(10);

        $count_cat = Advert::activeNotRecentlySold()
                          ->where('category', $cat->id)
                          ->count();

        return response()->json([
            'success' => true,
            'data' => $ads->items(),
            'category' => $cat,
            'total_count' => $count_cat,
            'pagination' => [
                'current_page' => $ads->currentPage(),
                'last_page' => $ads->lastPage(),
                'per_page' => $ads->perPage(),
                'total' => $ads->total(),
                'has_more' => $ads->hasMorePages()
            ]
        ]);
    }

    /**
     * Location + Subcategory search
     */
    private function locationSubcat($location, $slug)
    {
        $subcat = SubCategory::where('sub_cat_slug', $slug)->first();

        if (!$subcat) {
            return response()->json([
                'success' => false,
                'message' => 'Subcategory not found'
            ], 404);
        }

        $ads = Advert::with('firstImage')
                    ->activeNotRecentlySold()
                    ->where('sold', 'No')
                    ->where('state', $location)
                    ->where('sub_category', $subcat->id)
                    ->orderWithFeatured()
                    ->paginate(10);

        $count_subcat = Advert::activeNotRecentlySold()
                            ->where('sub_category', $subcat->id)
                            ->count();

        return response()->json([
            'success' => true,
            'data' => $ads->items(),
            'subcategory' => $subcat,
            'total_count' => $count_subcat,
            'pagination' => [
                'current_page' => $ads->currentPage(),
                'last_page' => $ads->lastPage(),
                'per_page' => $ads->perPage(),
                'total' => $ads->total(),
                'has_more' => $ads->hasMorePages()
            ]
        ]);
    }

    /**
     * Location + Brand search
     */
    private function locationBrand($location, $slug)
    {
        $brand = Brands::where('brand_slug', $slug)->first();

        if (!$brand) {
            return response()->json([
                'success' => false,
                'message' => 'Brand not found'
            ], 404);
        }

        $ads = Advert::with('firstImage')
                    ->activeNotRecentlySold()
                    ->where('state', $location)
                    ->where('brand', $brand->id)
                    ->orderWithFeatured()
                    ->paginate(10);

        $count_brand = Advert::activeNotRecentlySold()
                        ->where('brand', $brand->id)
                        ->count();

        return response()->json([
            'success' => true,
            'data' => $ads->items(),
            'brand' => $brand,
            'total_count' => $count_brand,
            'pagination' => [
                'current_page' => $ads->currentPage(),
                'last_page' => $ads->lastPage(),
                'per_page' => $ads->perPage(),
                'total' => $ads->total(),
                'has_more' => $ads->hasMorePages()
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/search/load-more",
     *     summary="Load more results for infinite scroll",
     *     tags={"Search"},
     *     @OA\Parameter(name="product", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="category", in="query", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="sub_category", in="query", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="brand", in="query", @OA\Schema(type="integer")),
     *     @OA\Parameter(name="location", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="buydirect", in="query", @OA\Schema(type="string")),
     *     @OA\Parameter(name="page", in="query", @OA\Schema(type="integer")),
     *     @OA\Response(response=200, description="More results")
     * )
     */
    public function loadMore(Request $request)
    {
        $query = Advert::with('firstImage')->activeNotRecentlySold();

        if ($request->filled('product')) {
            $query->where(function ($q) use ($request) {
                $q->where('ad_title', 'LIKE', '%' . $request->product . '%')
                  ->orWhere('ad_id', 'LIKE', '%' . $request->product . '%');
            });
        }

        if ($request->filled('category')) {
            $query->where('category', $request->category);
        }

        if ($request->filled('sub_category')) {
            $query->where('sub_category', $request->sub_category);
        }

        if ($request->filled('brand')) {
            $query->where('brand', $request->brand);
        }

        if ($request->filled('location')) {
            $query->where('state', $request->location);
        }

        if ($request->filled('buydirect')) {
            $query->where('buy_direct', $request->buydirect);
        }

        $ads = $query->orderWithFeatured()
                    ->paginate($request->input('per_page', 10));

        return response()->json([
            'success' => true,
            'data' => $ads->items(),
            'has_more' => $ads->hasMorePages(),
            'pagination' => [
                'current_page' => $ads->currentPage(),
                'last_page' => $ads->lastPage(),
                'per_page' => $ads->perPage(),
                'total' => $ads->total()
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/search/filters",
     *     summary="Get available filter options",
     *     tags={"Search"},
     *     @OA\Response(response=200, description="Filter options")
     * )
     */
    public function getFilters(Request $request)
    {
        return response()->json([
            'success' => true,
            'data' => [
                'categories' => Category::with('subCategories')->get(),
                'states' => State::all(),
                'price_ranges' => [
                    ['value' => 'under_20k', 'label' => 'Under ₦20,000'],
                    ['value' => '20k_120k', 'label' => '₦20,000 - ₦120,000'],
                    ['value' => '120k_1m', 'label' => '₦120,000 - ₦1,000,000'],
                    ['value' => '1m_10m', 'label' => '₦1,000,000 - ₦10,000,000'],
                    ['value' => 'above_10m', 'label' => 'Above ₦10,000,000'],
                ],
                'seller_types' => [
                    ['value' => 'all', 'label' => 'All Sellers'],
                    ['value' => 'yes', 'label' => 'Verified Sellers'],
                    ['value' => 'no', 'label' => 'Unverified Sellers'],
                ]
            ]
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/search/suggestions",
     *     summary="Get search suggestions",
     *     tags={"Search"},
     *     @OA\Parameter(name="q", in="query", required=true, @OA\Schema(type="string")),
     *     @OA\Response(response=200, description="Search suggestions")
     * )
     */
    public function getSuggestions(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'q' => 'required|string|min:2'
        ]);

        if ($validator->fails()) {
            return response()->json([
                'success' => false,
                'errors' => $validator->errors()
            ], 422);
        }

        $query = $request->input('q');

        $suggestions = Advert::where('ad_title', 'LIKE', '%' . $query . '%')
                            ->where('ad_status', 'active')
                            ->where('sold', 'No')
                            ->select('ad_title')
                            ->distinct()
                            ->limit(10)
                            ->pluck('ad_title');

        return response()->json([
            'success' => true,
            'data' => $suggestions
        ]);
    }
}