# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Marketplace Nigeria API Documentation'
  description: 'Complete REST API documentation for Marketplace Nigeria platform. This API allows you to manage adverts, users, messages, payments, and more.'
  version: '3.0'
servers:
  -
    url: 'http://127.0.0.1:8030'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can obtain your authentication token by logging in through the `/api/login` endpoint. The token should be included in the Authorization header as a Bearer token for all authenticated requests.'
security:
  -
    default: []
paths:
...
  /api/search:
    get:
      summary: ''
      operationId: getApiSearch
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Search
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                  description: 'Must be at least 3 characters.'
                  example: tyzptyixtefpmcizunvzpmiwuqenixrsztzmtjtvwhxfbgocfegcjbzmmol
                  nullable: true
                location:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                category:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                sub_category:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                brand:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                buydirect:
                  type: string
                  description: ''
                  example: null
                  nullable: true
      security: []
...
  '/api/search/location/{location}/{slug}':
    get:
      summary: ''
      operationId: getApiSearchLocationLocationSlug
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Search
      security: []
    parameters:
      -
        in: path
        name: location
        description: 'The location.'
        example: quaerat
        required: true
        schema:
          type: string
      -
        in: path
        name: slug
        description: 'The slug of the {location}.'
        example: eum
        required: true
        schema:
          type: string
...
  /api/search/suggestions:
    get:
      summary: ''
      operationId: getApiSearchSuggestions
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                q:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: xmfnhmpkudpvqwecuwnqpynbvebbnwupqbcwzxpotabyqtvwgtxsbnhfdxqwempdvghchtks
              required:
                - q
      security: []
...